+++ src/components/Highlighter/ReactElementHighlighter/index.ts
@@ -1,4 +1,5 @@
 export {
     ReactElementHighlighter,
     Props as ReactElementHighlighterProps,
+    ReactElementHighlighterTermResult,
 } from './ReactElementHighlighter';