+++ src/components/RadioButton/RadioButton.tsx
@@ -12,7 +12,7 @@ export interface Props extends React.HTMLProps<HTMLInputElement> {
     /** If the radio button should be disabled */
     disabled?: boolean;
     /** The label for the radio button */
-    children?: string;
+    children?: SingleReactNode;
 }
 
 const { block, elem } = bem('RadioButton', styles);
@@ -20,7 +20,7 @@ const { block, elem } = bem('RadioButton', styles);
 export const RadioButton = forwardRef<HTMLElement, Props>(
     ({ id, children, disabled = false, name, ...rest }, ref) => {
         return (
-            <div {...block({ ...rest })}>
+            <div {...block({ ...rest, disabled })}>
                 <input
                     {...rest}
                     {...elem('input')}